Update csharp-style.md
C# 7 was released in 2017 and after that also C# 8 (LTS released in 2019) and C# 10 (the latest LTS released earlier this month). 
May be you will find having sense to reduce the restriction for the syntax => for short methods definition. IMHO with advance of functional programming patterns inside of C#, this gem makes code more concise.